requirements.txt
numpy==1.16.4
torch==1.3.1
pillow==6.2.0
notebook==6.0.2
matplotlib==3.1.2
tensorflow==2.0.0
torchvision==0.4.2
tensorflow-privacy==0.2.2
